refactor(workstation): extract LogInkApp to workstation/runtime/app (#890 phase 5b)
The final phase 5 extraction. LogInkApp — the workstation's root React
component, ~1,940 lines of hooks, effects, callbacks, and state — moves
to src/workstation/runtime/app.ts. The 3 private helpers and the 3
context-loading helpers that only LogInkApp uses migrate with it:

  - predictNextFilter
  - computePromotedSelectionsSnapshot
  - enrichFilterActionWithRectification
  - safe (async error swallow)
  - loadLogInkContext (parallel overview fetch)
  - loadLogInkContextEntries (per-key boot-time loaders)

inkRuntime.ts shrinks to a small entry-point file — boot sequence only:
  1. Pick TTY vs non-TTY path
  2. Dynamically import the ESM-only `ink` + `react` modules
  3. Mount LogInkApp via React.createElement
  4. Install SIGTSTP / SIGCONT lifecycle handlers around the Ink instance
  5. Wait for exit

Phase 5 stats — all milestones since #890 opened:

  Phase 2 (#891)  Prune dead Inquirer-era branch                        -1,300
  Phase 3 (#894)  Promote git data layer to src/git/                    relocation
  Phase 4 (#893)  Promote chrome to src/workstation/chrome/             relocation
  Phase 8 (#895)  Workstation README                                    new doc
  Phase 5a.0 (#896) Foundation: types + visual helpers              inkRuntime -160
  Phase 5a.1 (#897) reflog, bisect, stash, worktrees                inkRuntime -394
  Phase 5a.2 (#898) compose, branches, tags, pullRequest            inkRuntime -486
  Phase 5a.3 (#899) status + conflicts                              inkRuntime -277
  Phase 5a.4 (#900) diff + split-diff helpers                       inkRuntime -417
  Phase 5a.5 (#901) history                                         inkRuntime -265
  Phase 5a.6 (#902) detail / inspector / preview                    inkRuntime -649
  Phase 5a.7 (#903) chrome + overlays + dispatchers                 inkRuntime -952
  Phase 5b   (this) LogInkApp + private helpers                     inkRuntime -2,307

inkRuntime.ts: 6,039 -> 132 LOC. -98% over phase 5.

src/workstation/ now has its final structural shape:

  workstation/
  ├── README.md
  ├── chrome/    (55 files — cross-cutting utilities)
  ├── runtime/   (11 files — LogInkApp + types + utils +
  │              splitDiff + promotedFilter + header + sidebar +
  │              footer + mainPanel + detailPanel + overlays)
  └── surfaces/  (13 dirs — 12 main surfaces + conflicts + detail family)

src/commands/log/ holds only the log command itself:
  - handler.ts / index.ts / config.ts / render.ts (log command core)
  - data.ts (log's row + commit-detail loader, used by both log and
    workstation — moves can happen later if needed)
  - interactive.ts / interactiveState.ts (non-TTY snapshot fallback)
  - commitCompose.ts (workstation compose state slice)
  - inkRuntime.ts (132 LOC — boot sequence only)
  - inkInput.ts, inkKeymap.ts, inkViewModel.ts, inkWorkflows.ts
    (workstation orchestration that phases 6/7 will split — out of
     scope for #890; tracked separately if/when prioritized)

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing — every check has been green at every step of phase 5.

No behavior change. The workstation boots, renders, navigates, and
handles input byte-identically before and after this PR. The
extraction is a pure code relocation; the React component graph is
unchanged.
